Fix bug with utils.js/getTimeDiff
When diff between now and publish time is less than 0.5s , the function used to return the previous unit instead of the first value of the next unit. E.g., if the diff was 59.5s, instead of returning 1m, it'd return 60s, as the if/else structure would deem that 59.5s is less than a minute, so it should be shown in seconds.